pyproject.toml
[tool.poetry]
name = "bn-adjustable-bed"
version = "0.0.0"
description = "BN Adjustable Bed Mock API and Socket Interface"
include = ["bn_adjustable_bed/config"]

homepage = "https://github.com/trevorlauder/bn-adjustable-bed"
repository = "https://github.com/trevorlauder/bn-adjustable-bed"

authors = [
    "Trevor Lauder <trevor@lauder.family>",
]

maintainers = [
    "Trevor Lauder <trevor@lauder.family>",
]

license = "MIT"

readme = "README.md"

classifiers = [
    "Topic :: Home Automation",
]

[tool.poetry.scripts]
bn-app-api = "bn_adjustable_bed.app_api:main"
bn-controller-api = "bn_adjustable_bed.controller_api:main"
bn-bed-socket = "bn_adjustable_bed.bed_socket:main"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.70"
redis = "^3.5"
pyyaml = "^6.0"
uvicorn = "^0.15"
pytest = { version = "^6.2", optional = true }
pytest-mock = { version = "^3.6", optional = true }
pytest-cov = { version = "^3.0", optional = true }
requests = { version = "^2.26", optional = true }

[tool.poetry.dev-dependencies]
tox = "^3.24"

[tool.poetry.extras]
test = ["pytest", "pytest-mock", "pytest-cov", "requests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = clean,py310,report
isolated_build = true

[testenv]
extras =
    test
commands =
    pytest --cov --cov-append --cov-report=term-missing {posargs:-vv}
depends =
    {py310}: clean
    report: py310

[testenv:clean]
extras =
    test
commands =
    coverage erase

[testenv:report]
extras =
    test
commands =
    coverage xml
    coverage report
"""

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
